# [libc] Speed up memmove overlapping check